+++ packages/render/src/browser/render.tsx
@@ -1,42 +1,7 @@
-import { convert } from 'html-to-text';
 import { Suspense } from 'react';
-import type { ReactDOMServerReadableStream } from 'react-dom/server';
-import { pretty } from '../node';
+import { pretty, toPlainText } from '../node';
 import type { Options } from '../shared/options';
-import { plainTextSelectors } from '../shared/plain-text-selectors';
-
-const decoder = new TextDecoder('utf-8');
-
-const readStream = async (stream: ReactDOMServerReadableStream) => {
-  const chunks: Uint8Array[] = [];
-
-  const writableStream = new WritableStream({
-    write(chunk: Uint8Array) {
-      chunks.push(chunk);
-    },
-    abort(reason) {
-      throw new Error('Stream aborted', {
-        cause: {
-          reason,
-        },
-      });
-    },
-  });
-  await stream.pipeTo(writableStream);
-
-  let length = 0;
-  chunks.forEach((item) => {
-    length += item.length;
-  });
-  const mergedChunks = new Uint8Array(length);
-  let offset = 0;
-  chunks.forEach((item) => {
-    mergedChunks.set(item, offset);
-    offset += item.length;
-  });
-
-  return decoder.decode(mergedChunks);
-};
+import { readStream } from '../shared/read-stream.browser';
 
 export const render = async (node: React.ReactNode, options?: Options) => {
   const suspendedElement = <Suspense>{node}</Suspense>;
@@ -51,6 +16,7 @@ export const render = async (node: React.ReactNode, options?: Options) => {
         onError(error: unknown) {
           reject(error);
         },
+        progressiveChunkSize: Number.POSITIVE_INFINITY,
       })
       .then(readStream)
       .then(resolve)
@@ -58,10 +24,7 @@ export const render = async (node: React.ReactNode, options?: Options) => {
   });
 
   if (options?.plainText) {
-    return convert(html, {
-      selectors: plainTextSelectors,
-      ...options.htmlToTextOptions,
-    });
+    return toPlainText(html, options.htmlToTextOptions);
   }
 
   const doctype =