+++ packages/render/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @react-email/render
 
+## 1.2.3
+
+### Patch Changes
+
+- 4e43552: use edge exports in convex runtime
+
+## 1.2.2
+
+### Patch Changes
+
+- 174f9cd: fix builds failing with esbuild and react-dom@18
+
+## 1.2.1
+
+### Patch Changes
+
+- 036c46c: fix MessageChannel errors on edge environments
+
+## 1.2.0
+
+### Minor Changes
+
+- fc2f407: add toPlainText utility and deprecate plainText option on `render`
+
+## 1.1.4
+
+### Patch Changes
+
+- f5f35f1: fix hydration markers on React canary/Next.js latest when rendering large email templates
+
 ## 1.1.3
 
 ### Patch Changes